// src/client/models/RatingSummary.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Describes one rating summary.
 */
export type RatingSummary = {
  /**
   * The id of the coffee
   */
  coffee_id: string;
  /**
   * Number of ratings for coffee
   */
  rating_count: number;
  /**
   * The current average rating for the coffee
   */
  rating_average: number;
};